plume model debug
Now appears to be producing something sensible for the MISMIP geometry. Have to set prefactor to 1.46 i.e.. double that in Favier 2019. I've used the ad hoc temperature scaling which results in melt rate more concentrated near GL, so this discrepancy is in the direction we expect.
Need to finish the script that produces the figure