fixed issue #35
Pickups adjusted so that the new simulation is endowed with both the old model and the old clock (i.e. don't update the model is default and non-optional). Retained option to update the output parameters, but now this default to using the new output parameters, which seems like the correct approach logically. Ideally, would use just the fields from first simulation, but this violated immutabiltiy of model. (is there are way around this?)